requirements.txt
aprslib==0.6.46
astroid==2.2.5
certifi==2019.9.11
chardet==3.0.4
Click==7.0
dnspython==1.16.0
httplib2==0.13.1
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.17.2
oauth2==1.9.0.post1
paho-mqtt==1.4.0
pylint==2.3.1
pymongo==3.9.0
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
retrying==1.3.3
six==1.12.0
typed-ast==1.4.0
tzlocal==2.0.0
urllib3==1.25.3
wrapt==1.11.2
